backend/pyproject.toml
[project]
name = "statesqft-backend"
version = "0.1.0"
description = "StateSqft API — Supabase JWT + OpenAI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
    "PyJWT[crypto]>=2.8.0",
    "openai>=1.12",
    "supabase>=2.4",
    "pydantic>=2.0",
    "pandas>=2.0",
    "statsmodels>=0.14",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
